requirements.txt
paho-mqtt==1.5.1
python-telegram-bot==12.8
